requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
arrow==1.3.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.8
distro==1.9.0
email-validator==2.1.0
exceptiongroup==1.2.2
filelock==3.13.4
Flask==3.0.0
fonttools==4.38.0
frozenlist==1.5.0
Frozen-Flask==1.0.1
fsspec==2024.10.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.4
litellm==1.52.11
MarkupSafe==2.1.5
matplotlib==3.7.0
multidict==6.1.0
nose==1.3.7
numpy==1.24.2
openai==1.54.5
packaging==23.0
Pillow==9.4.0
platformdirs==4.2.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sniffio==1.3.1
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
tryagain==1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.0
Werkzeug==3.0.1
wget==3.2
yarl==1.17.2
zipp==3.21.0
